refactor: Remove legacyAppLoader files and configurations
