+++ client/apps/legacyAppLoader/app.config.js
