+++ client/apps/legacyAppLoader/src/app-loader/components/AppModuleLoader.tsx
