+++ client/apps/legacyAppLoader/src/app-loader/components/MessagePage.tsx
