+++ client/apps/legacyAppLoader/src/app-loader/components/ProgressLoader.tsx
