+++ client/apps/legacyAppLoader/src/app-loader/components/index.ts
