+++ client/apps/legacyAppLoader/src/app-loader/legacy-interopt/LegacyAuthContainer.ts
