+++ client/apps/legacyAppLoader/src/app-loader/legacy-interopt/LegacyAuthUser.ts
