+++ client/apps/legacyAppLoader/src/app-loader/legacy-interopt/LegacyContextManager.ts
