+++ client/apps/legacyAppLoader/src/app-loader/legacy-interopt/components/create-legacy-context-component.tsx
