+++ client/apps/legacyAppLoader/src/app-loader/legacy-interopt/create-service-resolver.ts
