+++ client/apps/legacyAppLoader/src/app-loader/legacy-interopt/static.ts
