+++ client/apps/legacyAppLoader/src/app-loader/styles.ts
