+++ client/apps/legacyAppLoader/src/globalResources.ts
