+++ client/apps/legacyAppLoader/tsconfig.docker.json
