+++ client/apps/legacyAppLoader/tsconfig.json
