+++ client/apps/legacyAppLoader/vite.config.js
