+++ client/package.json
@@ -7,6 +7,7 @@
     "start": "npx lerna run serve",
     "build": "npx lerna run build",
     "build:apps": "npx lerna run build:apps",
+    "build:appLoader": "npx lerna run build:appLoader",
     "test:all": "test:coverage && build:coverage",
     "test": "lerna  run test",
     "test:coverage": "lerna run test:coverage",
